Splitting into a B2 app and a TPS app.  A bit messy in terms of naming conventions, but seems to work all right.  Next up is getting TPS challenge definitions squared away
